Fix approval request handling in Change creation form
