+++ templates/components/itilobject/fields/global_validation.html.twig
@@ -31,38 +31,47 @@
 
 {% import 'components/form/fields_macros.html.twig' as fields %}
 
-{% if item.isField('global_validation') %}
-   {% if item.isNewItem() %}
-      {% set validation_right = '' %}
-      {% if params['type'] == constant('Ticket::INCIDENT_TYPE') and has_profile_right('ticketvalidation', constant('TicketValidation::CREATEINCIDENT')) %}
+{% set validation = item.getValidationClassInstance() %}
+
+{% if validation is not null %}
+   {% set create_right_val = constant('CREATE') %}
+   {% set validation_right = 'validate' %}
+   {% if item.getType() == 'Ticket' %}
+      {% if params['type'] == constant('Ticket::INCIDENT_TYPE') %}
+         {% set create_right_val = constant('TicketValidation::CREATEINCIDENT') %}
          {% set validation_right = 'validate_incident' %}
-      {% endif %}
-      {% if params['type'] == constant('Ticket::DEMAND_TYPE') and has_profile_right('ticketvalidation', constant('TicketValidation::CREATEREQUEST')) %}
+      {% elseif params['type'] == constant('Ticket::DEMAND_TYPE') %}
+         {% set create_right_val = constant('TicketValidation::CREATEREQUEST') %}
          {% set validation_right = 'validate_request' %}
       {% endif %}
+   {% endif %}
+   {% set can_create = has_profile_right(static(validation, 'rightname'), create_right_val) %}
 
-      {% set required = {} %}
-      {% if field_options.fields_template.isMandatoryField('_add_validation') %}
-         {% set required = {
-            'required': true,
-         } %}
+   {% if item.isNewItem() %}
+      {% if can_create %}
+         {% set required = {} %}
+         {% if field_options.fields_template.isMandatoryField('_add_validation') %}
+            {% set required = {
+               'required': true,
+            } %}
+         {% endif %}
+
+         {{ fields.field(
+            '_add_validation',
+            call('TicketValidation::dropdownValidator', [{
+               'name': 'users_id_validate',
+               'entity': item.fields['entities_id'],
+               'right': validation_right,
+               'users_id_validate': params['users_id_validate'],
+               'display': false,
+               'disabled': (not canupdate),
+            }|merge(required)]),
+            __('Approval request'),
+            field_options
+         ) }}
+
+         <input type="hidden" name="_add_validation" value="{{ params['_add_validation'] }}" />
       {% endif %}
-
-      {{ fields.field(
-         '_add_validation',
-         call('TicketValidation::dropdownValidator', [{
-            'name': 'users_id_validate',
-            'entity': item.fields['entities_id'],
-            'right': validation_right,
-            'users_id_validate': params['users_id_validate'],
-            'display': false,
-            'disabled': (not canupdate),
-         }|merge(required)]),
-         __('Approval request'),
-         field_options
-      ) }}
-
-      <input type="hidden" name="_add_validation" value="{{ params['_add_validation'] }}" />
    {% else %}
         {% if not itiltemplate.isHiddenField('global_validation') %}
             {{ fields.htmlField(
